taint/test/mochitest/test_url_search_params.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Taint URL Search Params propagation</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>

      function check_tainted(str) {
          ok(str.taint.length > 0, "Check tainted");
      }

      function check_untainted(str) {
          ok(str.taint.length == 0, "Check untainted");
      }

      // Check untainted
      const paramsString = "q=URLUtils.searchParams&topic=api";
      const searchParams = new URLSearchParams(paramsString);

      check_untainted(searchParams.get("q"));
      check_untainted(searchParams.get("topic"));
      check_untainted(searchParams.toString());

      // Tainted basic checks
      const taintedParamsString = "tainted=" + String.tainted("hello")+ "&untainted=goodbye";
      const taintedSearchParams = new URLSearchParams(taintedParamsString);

      check_tainted(taintedSearchParams.get("tainted"));
      check_untainted(taintedSearchParams.get("untainted"));
      check_tainted(taintedSearchParams.toString());

      taintedSearchParams.delete("tainted");
      check_untainted(taintedSearchParams.toString());

    </script>
  </head>
  <body>
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <pre id="test"></pre>
  </body>
</html>
